trash cleaning

+++ srcs/main.cpp
@@ -14,9 +14,6 @@ int ft_socket_init(Server & server, int opt) {
 	addr.sin_port = htons(std::atoi(server.getPort().c_str()));
 	addr.sin_addr.s_addr = inet_addr(server.getHost().c_str());
 
-	// addr.sin_port = htons(50001);
-	// addr.sin_addr.s_addr = htonl(INADDR_ANY);
-
 	if (setsockopt(socket_fd, SOL_SOCKET, SO_REUSEADDR, &opt, sizeof(opt)) < 0)
 		std::cout << "SET_SOCK_OPT ERROR";
 	
@@ -60,20 +57,17 @@ bool ft_check_evlist_error(std::vector<struct kevent> & chlist, std::vector<stru
 		}
 		close(chlist[i].ident);
 		chlist.erase(chlist.begin() + i);
-        // evlist.clear();
-        // evlist.reserve(chlist.size());
-		// std::cout <<  chlist[i].ident << " " << evlist[0].ident << " Read direction of socket has shutdown\n";
 		return (true);
 	}
 	return (false);
 }
 
-bool ft_check_new_connection(unsigned int & socket_fd, int & i, std::vector<struct kevent> & chlist, \
+bool ft_check_new_connection(int & socket_fd, int & i, std::vector<struct kevent> & chlist, \
 								std::vector<struct kevent> & evlist, std::map<int, Client> & clients) {
 	struct sockaddr_in client;
     socklen_t client_size = sizeof(client);
 	
-	if (evlist[i].ident == socket_fd) {
+	if (evlist[i].ident == static_cast<unsigned int>(socket_fd)) {
 		int accept_fd = accept(socket_fd, (struct sockaddr *)&client, &client_size);
 		if (accept_fd < 0) std::cout << "ACCEPT ERROR\n"; else std::cout << "ACCEPT OK\n";
 		std::cout << "FCNTL "  << fcntl(accept_fd, F_SETFL, O_NONBLOCK) << std::endl;
@@ -96,7 +90,6 @@ void ft_check_clients(int & i, std::vector<struct kevent> & chlist, std::vector<
     (void)chlist;
     int fd = evlist[i].ident;
 	if (evlist[i].filter & EVFILT_READ) {
-			/* We have data from the client */
 			ret = recv(fd, &buf, BUFFER_SIZE, 0);
 			if (ret < 0) {
 				std::cout << "recv ERROR";
@@ -114,7 +107,7 @@ void ft_check_clients(int & i, std::vector<struct kevent> & chlist, std::vector<
     }
 }
 
-void ft_check_fds(int & nev, unsigned int & socket_fd, std::vector<struct kevent> & chlist, \
+void ft_check_fds(int & nev, int & socket_fd, std::vector<struct kevent> & chlist, \
 					std::vector<struct kevent> & evlist, std::map<int, Client> & clients) {
 	for (int i = 0; i < nev; i++) {
 		if (evlist[i].flags & EV_ERROR) { /* Report errors */
@@ -131,19 +124,13 @@ void ft_check_fds(int & nev, unsigned int & socket_fd, std::vector<struct kevent
 
 
 int main(int argc, char **argv) {
-	// int         				kq, nev;
-	// unsigned int				socket_fd;
-	// std::vector<struct kevent>	chlist(1);
-	// std::vector<struct kevent>	evlist(1);
-	// std::vector<Server>			servers;
-
 	int					kq, nev;
 	std::vector<Server>	servers;
 	
 	if (argc == 1) {ft_parse(servers, DEFAULT_CONF);}
 	else {ft_parse(servers, argv[1]);}
 	
-	std::vector<unsigned int>	socket_fd(servers.size());
+	std::vector<int>	socket_fd(servers.size());
 	std::vector<struct kevent>	chlist(servers.size()), evlist(servers.size());
 
 	kq = kqueue_init();
@@ -170,36 +157,3 @@ int main(int argc, char **argv) {
 	}
 	return (0);
 }
-
-// int main(int argc, char **argv) {
-// 	int         				kq, nev;
-// 	unsigned int				socket_fd;
-// 	std::vector<struct kevent>	chlist(1);
-// 	std::vector<struct kevent>	evlist(1);
-// 	std::vector<Server>			servers;
-	
-// 	if (argc == 1) {
-// 		ft_parse(servers, DEFAULT_CONF);
-// 	}
-// 	else
-// 		ft_parse(servers, argv[1]);
-	
-// 	socket_fd = ft_socket_init(servers[0], 1);
-// 	kq = kqueue_init(chlist, socket_fd);
-
-// 	while (1) {
-//         nev = kevent(kq, &chlist.front(), chlist.size(), &evlist.front(), chlist.size(), NULL);
-        
-// 		if (nev < 0) {
-//             std::cout << "kevent ERROR\n"; exit(EXIT_FAILURE);
-//         }
-//         else if (nev > 0) {
-// 			// std::cout << nev << " kevent OK\n";
-// 			if (ft_check_evlist_error(chlist, evlist)) {
-// 				continue ;
-// 			}
-// 			ft_check_fds(nev, socket_fd, chlist, evlist, servers[0].getClients());
-//         }
-// 	}
-// 	return (0);
-// }
